+++ imagemanager/src/main/java/com/grumpyshoe/module/imagemanager/impl/model/PermissionExplanation.kt
@@ -1,0 +1,3 @@
+package com.grumpyshoe.module.imagemanager.impl.model
+
+data class PermissionExplanation(val title: Int, val message: Int)